Item weigh changed to int and multiplied by 1000. Now unit in grams.

+++ Source/REquipmentLib/REquipmentMgrComponent.h
@@ -35,7 +35,7 @@ public:
 private:
    // To evade endless when:
    // OnStatusUpdated -> CalcWeight -> SetEffect -> OnStatusUpdated
-   float LastWeightMax = 0;
+   int32 LastWeightMax = 0;
 
 protected:
    UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Rade|Equipment")
